backend/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clerk-backend-api>=1.1.2",
    "fastapi[standard]>=0.115.3",
    "pydantic-settings>=2.6.0",
    "requests>=2.32.3",
    "sqlalchemy-libsql>=0.1.0",
    "sqlmodel>=0.0.22",
    "tenacity>=9.0.0",
]
